examples: type the birdwatcher community test helper as two 16-bit halves
The helper is documented as an RFC 1997 `asn:value` community, so it now
takes the two 16-bit halves as `u16` and widens them while composing the
32-bit value; a caller can no longer pass a value that would overlap the
ASN bits. The existing callers' literals all fit and infer the narrower
type unchanged.

+++ examples/birdwatcher-adapter/src/main.rs
@@ -2917,8 +2917,8 @@ mod tests {
     use clap::CommandFactory;
 
     /// `asn:value` as the RFC 1997 32-bit community.
-    fn community(asn: u32, value: u32) -> u32 {
-        (asn << 16) | value
+    fn community(asn: u16, value: u16) -> u32 {
+        (u32::from(asn) << 16) | u32::from(value)
     }
 
     fn token_file(name: &str, contents: &str) -> PathBuf {
